Update Ability_TPS_Crouch: Block other abilities while crouching
I added a switch to block Abilities while Crouch is running. For example, this can prevent the player from jumping or dashing while crouching.  This is mapped via a list so that the player can add abilities themselves. Or maybe some will like crouchjumps ;) Added the most common abilities in onvalidate.